updates
Made minor updates to the R script and LDEN.CSV file.

Added output files for the sample data from Kintigh's LDEN.exe program. For each data file there are four files of output: the .txt files contain the local density coefficients for radii of 1, 2, and 5 meters. The .out files have the counts  for the local neighborhoods of each point. The R script formats things differently, but if you run the code at the end of the script, lda_test should have the same local density coefficients as the .txt file that used the same data. Similarly, counts_test should have the same counts as the .out file that used the same data and radius. LDEN.exe expects the type designations to be numbers, so the AZ_A1020... output from LDEN.exe has types 1-5, which correspond to the types in the .csv in alphabetical order (1 = bead, 2 = ceramic, 3 = groundstone, 4 = projectile point, 5 = stone tool). The types should be in the same order in all the output files.